feat: wxlogin
feat: get/set user info

Revert "feat: get/set user info"

This reverts commit 0e9d3b921d1955572c34c3aeb9fb28e6f89d1f70.

Revert "Revert "feat: get/set user info""

This reverts commit b59ca7462f72eb41d34df0a19a5bee7fab059188.

feat: user service

+++ src/service/user/model/user_model.go
@@ -8,7 +8,7 @@ type User struct {
 	Username     string
 	Avatar       string
 	School       string
-	Grade        int
+	Grade        int64
 	Introduction string
-	PostCount    int
+	PostCount    int64
 }
